[Feat]: 카테고리바 추가 & [style]: 전체위치 조정
